13 feat 가게 리뷰 목록 조회 api 구현

+++ src/main/java/umc/web/dto/ReviewRequestDTO.java
@@ -13,6 +13,6 @@ public class ReviewRequestDTO {
         @NotBlank
         String body;
         @Min(1) @Max(5)
-        int score;
+        Float score;
     }
 }
